Back-end/requirements.txt
colorama==0.4.4
Flask==2.0.2
Flask_RESTful==0.3.9
numpy==1.22.1
pandas==1.4.0
python_dateutil==2.8.2
python_whois==0.7.3
requests==2.27.1
tldextract==3.2.1
whois==0.9.13
